Updated waypoints to type double
All intermediate waypoints are type double for the purpose of a finer
resolution for the EA.

+++ UTM_Monopoly/UTM_Monopoly/Waypoint.hpp
@@ -41,7 +41,7 @@ protected:
     
 public:
     
-    vector<int> waypoint_telm;
+    vector<double> waypoint_telm;
     
     void get_waypoint();
